Fix cleanup according to review comments
Remove the need for a special test which sets the default settings.
Instead we do this when we first get access to a new daemon management
interface. This requires that some of the early tests must be set to
cleanup = false (test_uninstall_app) because they are using a mullvad
management interface which stops working after the test runs. However
this is a trivial problem to solve and the tests will fail if you mess
this up. There is no important ordering requirement in relation to
settings anymore.

No longer run daemon related cleanup after tests that do not take the
management interface have finished. This is fine since they are not able
to mess up the daemon without the interface.

Tests that use the old management interface do not run cleanup with that
interface. This caused too much code bloat without any real purpose.
Right now we do not need to clean the old daemon for any reason and this
can be added in the future if it turns out we need it.

Increase the amount of resetting that the cleanup function does.