Add test attributes always_run and must_succeed
Tests can now be specified as always_run tests which will always run
regardless of what filters are provided to the test framework.
Test can also be specified as needing to succeed which means the test
framework will stop running if the test fails.

Both of these default to false which means that filtering is practical
and normal tests will run properly after installation and login has been
done. It also means that if some test is temporarily failing due to
external circumstances (such as some network service being down) will
not cause the entire run to halt.