Improve e2e-test missing error by adding context
Previously, if no "matching" e2e test was found for a
package-under-test, the error logging would not indicate that it was due
to the test-runner failing to find the e2e-teset binary. Instead, it
would just print "Could not find package ..", which is more general and
not that helpful in this case.