Implement a generic connection check strategy
As part of our automated tests, we want to assure that the internet is
reachable through the VPN tunnel. To verify connectivity, we can request
internet resources which are not the currently connected relay. Also,
try something more elaborate than a ping (i.e. check for HTTP `200` from
`am.i.mullvad.net`).

This is prone to network failures, which is why we retry each request 3
times before we consider it a failure.

Use the generic strategy to re-implement `geoip_lookup_with_retries`.