Add `TAG` argument to `ci-runtests.sh`
To reference any remote git tag in the `ci-runtests.sh` script, the new
environment variable `TAG` was introduced. The `TAG` is any glob
pattern which can uniquely identify a valid git tag on the remote
instance. `main` will be used the default value for `TAG` if no other
is given.