Fix race condition in `test_revoked_device`
Wait for device validity cache to time out before checking if the device
is valid/invalid. This seemingly fixes the issue with
`test_revoked_device` sometimes failing on some platforms.

Note: We should probably look for a long-term fix. One suggestions is to
add daemon-function(s) for bypassing/invalidating cached values, which would be
useful in testing.